requirements.txt
fastapi
uvicorn
python-docx
langchain-google-genai
langchain-core
jinja2
pydantic
